# Conflicts:
#	src/main/java/ee/coffee/coffeeshop/controllers/OrderController.java
#	src/main/java/ee/coffee/coffeeshop/models/enums/OrderStatus.java
#	src/main/resources/templates/products.ftlh

+++ src/main/java/ee/coffee/coffeeshop/repositories/OrderRepository.java
@@ -4,12 +4,9 @@ import ee.coffee.coffeeshop.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
-
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
     Optional<Order> findById(Long id);
 }
